[TASK] Add Pseudo code to Set Port and Attenuation of WSS module
Add pseudo code for the function to configure multiple channels in
add/drop ports and apply attenuation to them. The function uses
long command format and forms the command structure dynamically
based on number of channels.

Add logic to convert attenuation values from float to integer which
is used for including in payload. Include logic to form the command
structure dynamically based on channels count.