[TASK] Add document on Assign Waveplan ID
Add document that explains about the purpose of the function
and when it is used. Add details about the arguments that function
takes, list of elements in command structure and how the command
structure is formed is explained.

Add required code snippets so that it contains the logic for
how the function is implemented. Mention about command structure
elements and number of bytes that each take with example. Explain
how command is passed to module and response is got back from it.
Add information for validation of response and function return.