[TASK] Add document on Set Channel Port & Attenuation in WSS
Explain the significance of configuration in WSS and how it is
configured. Give attenuation range and block value for the WSS
module. Add details about the arguments that function takes,
list of elements in command structure, maximum bytes that the
Long Command structure takes and how the command structure is
formed.

Provide value for each byte such as MID, CMD, OBJ ID, etc,.
so that the reader can understand when and where those values
should be used. Add logic for forming the command structure
dynamically based onn number of channels.

Add required code snippets so that it contains the view of
code how it actually looks like. Explain about start and stop
bytes and checksum calculation. Mention the UART transmit/receive
functions through which the command is sent and response is received.