[TASK] Add document on Set Channel Port & Attenuation in Twin WSS
This functon is used to set multiple channel port and attenuation
in both WSS in module. Each WSS has 20 ports and so with this
function, the port numbers can be considered from 1 to total number
of ports in both modules i.e., 1-40. Use of WSS ID and Altconf ID
is not required for this function as we are referring to both the
modules.

Add details about the arguments that function takes, list of elements
in command structure, maximum bytes that the Long Command structure
takes and how the command structure is formed.

Provide value for each byte such as MID, CMD, OBJ ID, etc,.
so that the reader can understand when and where those values
should be used. Add logic for forming the command structure
dynamically based onn number of channels.

Add required code snippets so that it contains the view of
code how it actually looks like. Explain about start and stop
bytes and checksum calculation. Mention the UART transmit/receive
functions through which the command is sent and response is received.